#Issue 1: Create an async version of the SecurityLogic interface
The existing pac4j SecurityLogic interface is heavily focused on synchronous logic. Create a new version of the interface which reflects the need to perform blocking tasks off the event loop thread and handle the results when they arrive, so that the SecurityLogic call itself can be handled from the eventloop thread.